vulkan: Rename vulkan files with prefix
