+++ framework/decode/CMakeLists.txt
@@ -66,8 +66,6 @@ target_sources(gfxrecon_decode
                     ${CMAKE_CURRENT_LIST_DIR}/json_writer.cpp
                     ${CMAKE_CURRENT_LIST_DIR}/decode_json_util.h
                     ${CMAKE_CURRENT_LIST_DIR}/decode_json_util.cpp
-                    ${CMAKE_CURRENT_LIST_DIR}/pnext_node.h
-                    ${CMAKE_CURRENT_LIST_DIR}/pnext_typed_node.h
                     ${CMAKE_CURRENT_LIST_DIR}/pointer_decoder_base.h
                     ${CMAKE_CURRENT_LIST_DIR}/pointer_decoder.h
                     ${CMAKE_CURRENT_LIST_DIR}/portability.h
@@ -99,6 +97,8 @@ target_sources(gfxrecon_decode
                     $<$<BOOL:${GFXRECON_TOCPP_SUPPORT}>:${CMAKE_CURRENT_LIST_DIR}/vulkan_cpp_util_datapack.h>
                     $<$<BOOL:${GFXRECON_TOCPP_SUPPORT}>:${CMAKE_CURRENT_LIST_DIR}/vulkan_cpp_util_datapack.cpp>
                     ${CMAKE_CURRENT_LIST_DIR}/vulkan_consumer_base.h
+                    ${CMAKE_CURRENT_LIST_DIR}/vulkan_pnext_node.h
+                    ${CMAKE_CURRENT_LIST_DIR}/vulkan_pnext_typed_node.h
                     ${CMAKE_CURRENT_LIST_DIR}/vulkan_stats_consumer.h
                     ${CMAKE_CURRENT_LIST_DIR}/stat_decoder_base.h
                     ${CMAKE_CURRENT_LIST_DIR}/stat_decoder_base.cpp
