+++ framework/decode/custom_vulkan_struct_decoders.h
@@ -28,9 +28,9 @@
 #include "decode/custom_vulkan_struct_decoders_forward.h"
 #include "decode/handle_pointer_decoder.h"
 #include "decode/pointer_decoder.h"
-#include "decode/pnext_node.h"
 #include "decode/string_decoder.h"
 #include "decode/struct_pointer_decoder.h"
+#include "decode/vulkan_pnext_node.h"
 #include "generated/generated_vulkan_struct_decoders_forward.h"
 #include "util/defines.h"
